INTERLOK-4160 Add new annotation so the UI can load the existing variables and properties in the work unit
Also change the way the workunit: url works to avoid relaying too much on the jar name inside the work unit (e.g. when using xml transform service using an xsl file inside the work unit)
On init the wok unit will register its id, found in the adapter-version artifactId or the default jar name and register it into a work unit registry.
When running if a workunit:workUnitName url is found, the url handler will find the right url to the work unit jar with the given name/id and find the file with in it.

This will help as the person in charge of implementing the work unit will not need to know the name of the jar file at run time.
If there are two work unit have the same id only the first one will be registered.